🛠️: refactor `FormBoolean` test to typescript
